% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACL.r
\docType{data}
\name{ACL}
\alias{ACL}
\title{ACL Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ACL
}
\description{
ACL Class
}
\section{Fields}{

\describe{
\item{\code{execute}}{can execute}

\item{\code{read}}{can read}

\item{\code{write}}{can write}
}}

\keyword{datasets}
