% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agave.r
\docType{data}
\name{Agave}
\alias{Agave}
\title{Agave API parent class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
api = Agave$new(baseUrl="https://public.agaveapi.org",
                username="myusername",
                password="mypassword")
}
\description{
Agave Class

Singleton utility class to ease auth, api client instantiation, and
session management.

Note: This class is NOT created as part of the R Swagger client
generation.
}
\details{
Create an Agave object pointing to your tenant:\preformatted{library(rAgave)
api = Agave$new(baseUrl="https://public.agaveapi.org",
                username="myusername",
                password="mypassword")
}

Once the object is instantiated, interact with it according to the
methods in the API documentation.

For example, create a new client with:\preformatted{clientData <- Client$new(clientName="my_client")
> api$clients$add_client(body=clientData)$content
}

You may also pass in a list, if preferred, over the object model\preformatted{api$clients$create(body=list(clientName="my_client"))
}

Access any endpoint with:\preformatted{api$systems$list_systems()
api$jobs$submit_job(body=Job$new(appId="head-1.0u1"))
}

Once a client is created, it is used by default to access the API.

To make use of an existing client, pass the client's credentials into the Agave constructor:\preformatted{api = Agave$new(baseUrl="https://public.agaveapi.co",
                username="myusername", password="mypassword",
                clientKey="123", clientSecret="abc")
}

Alternatively, the SDK will attempt to recover the client credentials from a predictable location disk. The default cache directory is \code{$HOME/.agave}. You can configure the location of the cache directory when instantiating a new \code{Agave} object:\preformatted{api = Agave$new(baseUrl="https://public.agaveapi.co",
                username="myusername", password="mypassword",
                cacheDir="/Dropbox/agave/cache")
}

The format of the cache file is recognized by all Agave SDK, so you can freely move across multiple SDK without having to reauthenticate each time.

Your password will never be written to the cache file.

rAgave will automatically update the cache file any time you create or update your client or authenticate. Thus, subsequent instantiations can be streamlined to use the no-arg constructor as rAgave will automatically refresh your access token for you.\preformatted{api = Agave$new()
}

rAgave leverages the \href{https://github.com/zatonovo/futile.logger}{futile.logger} package to provide a custom logging output similar to log4j. You can enabled logging at runtime by specifying the \code{logLevel} parameter in the constructor. The following log levels are predefined as constants: TRACE, DEBUG, INFO, WARN, FATAL. The log level is set to FATAL by default.\preformatted{api = Agave$new(logLevel = INFO)
}

You can also change the log level at runtime by upating the \code{logLevel} attibute of an Agave class instance.\preformatted{api = Agave$new()
api$logLevel <- TRACE
profile <- api$profiles$getDetails()
api$logLevel <- DEBUG
profile <- api$jobs$listJobs()
api$logLevel <- FATAL
}

When enabled, logs will be written to \emph{\strong{agave.log}} in the currrent directory (\code{getwd()}). You can specify an alternative location by specifying the \code{logFilePath} parameter in the constructor.\preformatted{api = Agave$new(logFilePath = "/var/log/rAgave.log", logLevel = INFO)
}
}
\section{Fields}{

\describe{
\item{\code{apps}}{instace of \link[rAgave:AppsAPI]{rAgave::AppsAPI}preconfigured to interact with the current auth configuration.

See \link[rAgave:AppsAPI]{rAgave::AppsApi}, \link[rAgave:Application]{rAgave::Application}

\strong{Usage}\preformatted{api$apps$listApps(limit=10)
api$apps$getAppDetails(id="fork-0.1.0", responseType="raw", pretty=TRUE)
}}

\item{\code{clients}}{\link[rAgave:ClientsApi]{rAgave::ClientsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:ClientsApi]{rAgave::ClientsApi}, \link[rAgave:Client]{rAgave::Client}

\strong{Usage}\preformatted{api$clients$listClients()
api$clients$addClient(clientName="my_client")
}}

\item{\code{jobs}}{\link[rAgave:JobsApi]{rAgave::JobsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:JobsApi]{rAgave::JobsApi}, \link{Job}

\strong{Usage}\preformatted{api$jobs$listJobs(limit=10)
api$jobs$getJobDetails(id="4083976443515834856-242ac113-0001-007", responseType="dt")
}}

\item{\code{files}}{\link[rAgave:FilesApi]{rAgave::FilesApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:FilesApi]{rAgave::FilesApi}, \link[rAgave:FileItem]{rAgave::FileItem}

\strong{Usage}\preformatted{api$files$list(path=".", systemId="linux.example.com", responseType="dt")
api$files$download(path="agave.png", systemId="linux.example.com", filename="agave-logo.png")
api$files$upload(destSystemId="linux.example.com", destFilePath=".", filename="agave-logo.png")
}}

\item{\code{meta}}{\link[rAgave:MetaApi]{rAgave::MetaApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:MetaApi]{rAgave::MetaApi}, \link[rAgave:Metadata]{rAgave::Metadata}, \link[rAgave:MetadataSchema]{rAgave::MetadataSchema}

\strong{Usage}\preformatted{api$meta$listMetadata(q=list("value.history.date"="2012-01-01T01:11:06-0600), limit=50)
}}

\item{\code{monitors}}{\link[rAgave:MonitorsApi]{rAgave::MonitorsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:MonitorsApi]{rAgave::MonitorsApi}, \link[rAgave:Monitor]{rAgave::Monitor}

\strong{Usage}\preformatted{api$monitors$listMonitors(limit=10)
api$monitors$getMonitorDetails(id="3725843997056232985-242ac11e-0001-014")
}}

\item{\code{notifications}}{\link[rAgave:NotificationsApi]{rAgave::NotificationsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:NotificationsApi]{rAgave::NotificationsApi}, \link[rAgave:Notification]{rAgave::Notification}

\strong{Usage}\preformatted{api$notifications$listNotifications(limit=10)
api$notifications$getNotificationDetails(id="5952569791667432985-242ac119-0001-011")
}}

\item{\code{profiles}}{\link[rAgave:ProfilesApi]{rAgave::ProfilesApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:ProfilesApi]{rAgave::ProfilesApi}, \link[rAgave:Profile]{rAgave::Profile}

\strong{Usage}\preformatted{api$profiles$listProfiles(limit=10)
api$profiles$getProfile()
api$profiles$getProfile(username="nryan")
}}

\item{\code{systems}}{\link[rAgave:SystemsApi]{rAgave::SystemsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:SystemsApi]{rAgave::SystemsApi}, \link[rAgave:Systems]{rAgave::Systems}

\strong{Usage}\preformatted{api$systems$listSystems(limit=10)
api$systems$getSystemDetails(id="compute.example.com")
}}

\item{\code{tags}}{\link[rAgave:TagApi]{rAgave::TagApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:TagApi]{rAgave::TagApi}, \link[rAgave:Tag]{rAgave::Tag}

\strong{Usage}\preformatted{api$apps$listTags(limit=10)
api$apps$getTagDetails(id="docker", responseType="raw", pretty=TRUE)
}}

\item{\code{tenants}}{\link[rAgave:TenantsApi]{rAgave::TenantsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:TenantsApi]{rAgave::TenantsApi}, \link[rAgave:Tenant]{rAgave::Tenant}

\strong{Usage}\preformatted{api$tenants$listTenants()
api$tenants$getTenantDetails(id="agave.prod")
}}

\item{\code{tokens}}{\link[rAgave:TokensApi]{rAgave::TokensApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:TokensApi]{rAgave::TokensApi}, \link[rAgave:Tokens]{rAgave::Tokens}

\strong{Usage}\preformatted{api$tokens$create()
api$tokens$refresh()
}}

\item{\code{transforms}}{\link[rAgave:TransformsApi]{rAgave::TransformsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:TransformsApi]{rAgave::TransformsApi}, \link[rAgave:Transform]{rAgave::Transform}

\strong{Usage}\preformatted{api$transforms$listTransforms(limit=10)
api$transforms$getTransformDetails(id="zip-1.0")
}}

\item{\code{uuids}}{\link[rAgave:UuidsApi]{rAgave::UuidsApi} preconfigured to interact with the current auth configuration.

See \link[rAgave:UuidsApi]{rAgave::UuidsApi}, \link[rAgave:Uuid]{rAgave::Uuid}

\strong{Usage}\preformatted{api$uuids$lookup(uuid="4083976443515834856-242ac113-0001-007")
api$uuids$resolve(uuid="4083976443515834856-242ac113-0001-007")
}}

\item{\code{cache}}{Instance of the [rAgave::AgaveCache currently in use}

\item{\code{tokenInfo}}{The current \link[rAgave:Token]{rAgave::Token} object}

\item{\code{clientInfo}}{The active \link[rAgave:Client]{rAgave::Client} application}

\item{\code{tenantInfo}}{The active \link[rAgave:Tenant]{rAgave::Tenant} object`}
}}

\section{ManualMethods}{

\describe{

\code{$store()} persists the current token, tenant, and client info to disk in the auth cache file.

\strong{Usage}\preformatted{api$store()
}

\code{$restore(client, token, tenant)} refreshes the client, token and tenant currently being used by the instance by reloading If any parameters are passed into the call, they will override the current values and be persisted to disk.

\strong{Usage}\preformatted{api$restore()
api$restore(token=myToken)
}

}
}

\seealso{
\link[rAgave:ClientsAPI]{rAgave::ClientsAPI}, \link[rAgave:JobsAPI]{rAgave::JobsAPI}, \link[rAgave:FilesAPI]{rAgave::FilesAPI}, \link[rAgave:MetaAPI]{rAgave::MetaAPI}, \link[rAgave:MonitorsAPI]{rAgave::MonitorsAPI}, \link[rAgave:NotificationsAPI]{rAgave::NotificationsAPI}, \link[rAgave:ProfilesAPI]{rAgave::ProfilesAPI}, \link[rAgave:SystemsAPI]{rAgave::SystemsAPI}, \link[rAgave:TagAPI]{rAgave::TagAPI}, \link[rAgave:TenantsAPI]{rAgave::TenantsAPI}, \link[rAgave:TokensAPI]{rAgave::TokensAPI}, \link[rAgave:TransformsAPI]{rAgave::TransformsAPI}, \link[rAgave:UuidsApi]{rAgave::UuidsApi}, \link[rAgave:AgaveCache]{rAgave::AgaveCache}, \link[rAgave:Token]{rAgave::Token}, \link[rAgave:Client]{rAgave::Client}, \link[rAgave:Tenant]{rAgave::Tenant}
}
\keyword{datasets}
