% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgaveCache.r
\docType{data}
\name{AgaveCache}
\alias{AgaveCache}
\title{Agave Cache utility class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgaveCache
}
\description{
rAgave.AgaveCache
}
\details{
AgaveCache Class

Handles access to the authentication cache including reading,
writing from default system locations.

NOTE: This class is NOT generated by the swagger code generator program.
}
\section{Methods}{

\describe{

load

write

getClient

setClient

getToken

setToken

getTenant

setTenant

getProperty

}
}

\keyword{datasets}
