% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgaveNullResponse.r
\docType{data}
\name{AgaveNullResponse}
\alias{AgaveNullResponse}
\title{AgaveNullResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgaveNullResponse
}
\description{
Null response from the API
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{status}}{success or failure}

\item{\code{version}}{The current api version}
}}

\keyword{datasets}
