% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgaveResponse.r
\docType{data}
\name{AgaveResponse}
\alias{AgaveResponse}
\title{AgaveResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgaveResponse
}
\description{
This is the wrapper within which all API responses are returned.
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{status}}{success or failure}

\item{\code{version}}{The current api version}
}}

\keyword{datasets}
