% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.r
\docType{data}
\name{Application}
\alias{Application}
\title{Application Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Application
}
\description{
Applications are a self-describing representation of an executable software item. See the <a href=&quot;http://agaveapi.co/documentation/tutorials/app-management-tutorial/&quot;>App Management Tutorial</a> for more information.
}
\section{Fields}{

\describe{
\item{\code{available}}{Whether the application is available.}

\item{\code{checkpointable}}{Whether the application supports checkpointing.}

\item{\code{defaultMaxRunTime}}{The max execution time that should be used if none is given in a job description. Ignore if the system does not support schedulers.}

\item{\code{defaultMemoryPerNode}}{The default memory in GB to pass to the scheduler if none is given in the job description. This must be less than the max memory parameter in the target queue definition.}

\item{\code{defaultNodeCount}}{The number of nodes that should be used if none is given in a job description. Ignore if the system does not support schedulers.}

\item{\code{defaultProcessorsPerNode}}{The number of processors to pass to the scheduler if none are given in the job description. This must be 1 if the app is serial.}

\item{\code{defaultQueue}}{The queue on the execution system that should be used if none is given in a job description. Ignore if the system does not support schedulers.}

\item{\code{deploymentPath}}{The location in default storage system containing the application wrapper and dependencies of the user}

\item{\code{deploymentSystem}}{The system id of the storage system where this app should run.}

\item{\code{executionSystem}}{The system id of the execution system where this app should run.}

\item{\code{executionType}}{The execution type of the application. If you're unsure, it's probably HPC.}

\item{\code{helpURI}}{The URL where users can go for more information about the app.}

\item{\code{id}}{Unique id of this app. Comprised of the app name-version.}

\item{\code{icon}}{The icon to associate with this app.}

\item{\code{inputs}}{The inputs files for this application.}

\item{\code{label}}{The label to use when generating forms.}

\item{\code{lastModified}}{The date this application was last modified in ISO 8601 format.}

\item{\code{longDescription}}{The full text description of this input to use when generating forms.}

\item{\code{modules}}{An array of modules to load prior to the execution of the application.}

\item{\code{name}}{The name of the application. The name does not have to be unique, but the combination of name and version does.}

\item{\code{ontology}}{An array of ontology values describing this application.}

\item{\code{outputs}}{The outputs files for this application.}

\item{\code{parallelism}}{The parallelism type of the application. If you're unsure, it's probably SERIAL.}

\item{\code{parameters}}{The inputs parameters for this application.}

\item{\code{isPublic}}{Whether the application is public or private.}

\item{\code{revision}}{The number of times this application has been revised.}

\item{\code{shortDescription}}{The short description of this application.}

\item{\code{tags}}{An array of tags related to this application.}

\item{\code{templatePath}}{The path to the wrapper script relative to the deploymentPath.}

\item{\code{testPath}}{The path to the test script relative to the deploymentPath.}

\item{\code{uuid}}{The UUID of this application. UUID are 36 alphanumeric string.}

\item{\code{version}}{The version of the application in #.#.# format. While the version does not need to be unique, the combination of name and version does have to be unique.}
}}

\keyword{datasets}
