% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationAction.r
\docType{data}
\name{ApplicationAction}
\alias{ApplicationAction}
\title{ApplicationAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationAction
}
\description{
Generic action to take on an app
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
