% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationActionType.r
\docType{data}
\name{ApplicationActionType}
\alias{ApplicationActionType}
\title{ApplicationActionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationActionType
}
\description{
ApplicationActionType Class
}
\keyword{datasets}
