% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationArgument.r
\docType{data}
\name{ApplicationArgument}
\alias{ApplicationArgument}
\title{ApplicationArgument Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationArgument
}
\description{
This is the base model for all Application input, output, and parameter objects.
}
\section{Fields}{

\describe{
\item{\code{details}}{}

\item{\code{id}}{The id of this argument. This will be the replacement string in your wrapper scripts.}

\item{\code{semantics}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
