% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationArgumentDataSemantics.r
\docType{data}
\name{ApplicationArgumentDataSemantics}
\alias{ApplicationArgumentDataSemantics}
\title{ApplicationArgumentDataSemantics Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationArgumentDataSemantics
}
\description{
Describes the semantic definition of this inputs and the filetypes it represents. Multiple ontologies and values are supported.
}
\section{Fields}{

\describe{
\item{\code{fileTypes}}{The file types acceptable for this output.}

\item{\code{maxCardinality}}{The maximum number of times this output may appear.}

\item{\code{minCardinality}}{The minimum number of times this output may appear. -1 is unlimited}

\item{\code{ontology}}{}
}}

\keyword{datasets}
