% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationArgumentDetails.r
\docType{data}
\name{ApplicationArgumentDetails}
\alias{ApplicationArgumentDetails}
\title{ApplicationArgumentDetails Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationArgumentDetails
}
\description{
ApplicationArgumentDetails Class
}
\section{Fields}{

\describe{
\item{\code{description}}{Description of this input.}

\item{\code{label}}{The label for this input}

\item{\code{argument}}{The command line value of this input (ex -n, --name, -name, etc)}

\item{\code{showArgument}}{Whether the argument value should be passed into the wrapper at run time}

\item{\code{repeatArgument}}{Whether the argument value should be repeated in front of each user-supplied input before injection into the wrapper template at runtime}
}}

\keyword{datasets}
