% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationArgumentSemantics.r
\docType{data}
\name{ApplicationArgumentSemantics}
\alias{ApplicationArgumentSemantics}
\title{ApplicationArgumentSemantics Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationArgumentSemantics
}
\description{
Describes the semantic definition of this input/output and the filetypes it represents. Multiple ontologies and values are supported. <a href=&quot;http://agaveapi.co/documentation/tutorials/app-management-tutorial/#app-outputs&quot;>Read more.</a>
}
\section{Fields}{

\describe{
\item{\code{fileTypes}}{The file types acceptable for this output.}

\item{\code{maxCardinality}}{The maximum number of times this output may appear.}

\item{\code{minCardinality}}{The minimum number of times this output may appear. -1 is unlimited}

\item{\code{ontology}}{}
}}

\keyword{datasets}
