% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationArgumentValue.r
\docType{data}
\name{ApplicationArgumentValue}
\alias{ApplicationArgumentValue}
\title{ApplicationArgumentValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationArgumentValue
}
\description{
Describes the validation, default value, and primary types associated with this ApplicationArgument
}
\section{Fields}{

\describe{
\item{\code{defaultValue}}{The default value for this parameter. The type will be determined by the value.type field.}

\item{\code{enquote}}{Whether the argument value should be surrounded by quotation marks before injection into the wrapper template at runtime}

\item{\code{order}}{The order in which this parameter should be printed when generating an execution command for forked execution. This will also be the order in which paramters are returned in the response json.}

\item{\code{required}}{Is this parameter required? If visible is false, this must be true.}

\item{\code{validator}}{The regular expression used to validate this parameter value. For enumerations, separate values with |}

\item{\code{visible}}{Should this parameter be visible? If not, there must be a default and it will be required.}
}}

\keyword{datasets}
