% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCloneAction.r
\docType{data}
\name{ApplicationCloneAction}
\alias{ApplicationCloneAction}
\title{ApplicationCloneAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationCloneAction
}
\description{
Clones an existing app for private use.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
