% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationDisableAction.r
\docType{data}
\name{ApplicationDisableAction}
\alias{ApplicationDisableAction}
\title{ApplicationDisableAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationDisableAction
}
\description{
Disables an app from use.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
