% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationEnableAction.r
\docType{data}
\name{ApplicationEnableAction}
\alias{ApplicationEnableAction}
\title{ApplicationEnableAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationEnableAction
}
\description{
Make a disabled application available once again.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
