% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationExecutionType.r
\docType{data}
\name{ApplicationExecutionType}
\alias{ApplicationExecutionType}
\title{ApplicationExecutionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationExecutionType
}
\description{
ApplicationExecutionType Class
}
\keyword{datasets}
