% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationInput.r
\docType{data}
\name{ApplicationInput}
\alias{ApplicationInput}
\title{ApplicationInput Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationInput
}
\description{
App inputs describe data inputs supported by your app. Each input can represent one or more files or folders. Inptus can be optional or required and may physically reside anywhere accessible using any of the data protocols supported by Agave.  <a href=&quot;http://agaveapi.co/documentation/tutorials/app-management-tutorial/#app-inputs&quot;>Read more.</a>
}
\section{Fields}{

\describe{
\item{\code{details}}{}

\item{\code{id}}{The id of this argument. This will be the replacement string in your wrapper scripts.}

\item{\code{semantics}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
