% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationOutput.r
\docType{data}
\name{ApplicationOutput}
\alias{ApplicationOutput}
\title{ApplicationOutput Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationOutput
}
\description{
In addition to describing the inputs and parameters that your wrapper script requires, it is often helpful to provide the expected outputs when running your app. This is the purpose of the outputs attribute. App outputs specify an array of JSON objects describing the data that should be present when your app completes. It is entirely optional and provided, at this point in time, for reference purpose only.  <a href=&quot;http://agaveapi.co/documentation/tutorials/app-management-tutorial/#app-outputs&quot;>Read more.</a>
}
\section{Fields}{

\describe{
\item{\code{details}}{}

\item{\code{id}}{The id of this argument. This will be the replacement string in your wrapper scripts.}

\item{\code{semantics}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
