% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationParameter.r
\docType{data}
\name{ApplicationParameter}
\alias{ApplicationParameter}
\title{ApplicationParameter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationParameter
}
\description{
App parameters define the command-line arguments (flags, arguments, etc.) needed by your wrapper script in order to properly run your application code. <a href=&quot;http://agaveapi.co/documentation/tutorials/app-management-tutorial/#app-parameters&quot;>Read more.</a>
}
\section{Fields}{

\describe{
\item{\code{details}}{}

\item{\code{id}}{The id of this argument. This will be the replacement string in your wrapper scripts.}

\item{\code{semantics}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
