% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationParameterEnumValueItem.r
\docType{data}
\name{ApplicationParameterEnumValueItem}
\alias{ApplicationParameterEnumValueItem}
\title{ApplicationParameterEnumValueItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationParameterEnumValueItem
}
\description{
Enumerated single attribute object representing an enumeration value for an ApplicationParamaeterEnumValue. While a single pimary type may be specified, an ApplicationParameterEnumValueItem is the recommended structure as it allows for better display for consumers rendering enumerated values as HTML select elemetns.
}
\section{Fields}{

\describe{
\item{\code{key}}{The enumerated value.}

\item{\code{value}}{The value of the enumeration to display in a dropdown list}
}}

\keyword{datasets}
