% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationParameterNumericValue.r
\docType{data}
\name{ApplicationParameterNumericValue}
\alias{ApplicationParameterNumericValue}
\title{ApplicationParameterNumericValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationParameterNumericValue
}
\description{
An ApplicationParameter representing a numeric value
}
\section{Fields}{

\describe{
\item{\code{defaultValue}}{The default value for this parameter. The type will be determined by the value.type field.}

\item{\code{enumValues}}{An array of enumerated object values.}

\item{\code{enquote}}{Whether the argument value should be surrounded by quotation marks before injection into the wrapper template at runtime}

\item{\code{order}}{The order in which this parameter should be printed when generating an execution command for forked execution. This will also be the order in which paramters are returned in the response json.}

\item{\code{required}}{Is this parameter required? If visible is false, this must be true.}

\item{\code{type}}{The type of this parameter value. (Acceptable values are: &quot;string&quot;, &quot;number&quot;, &quot;enumeration&quot;, &quot;bool&quot;, &quot;flag&quot;)}

\item{\code{validator}}{The regular expression used to validate this parameter value. For enumerations, separate values with |}

\item{\code{visible}}{Should this parameter be visible? If not, there must be a default and it will be required.}
}}

\keyword{datasets}
