% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationParameterType.r
\docType{data}
\name{ApplicationParameterType}
\alias{ApplicationParameterType}
\title{ApplicationParameterType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationParameterType
}
\description{
The primary type of the application
}
\keyword{datasets}
