% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationPublishAction.r
\docType{data}
\name{ApplicationPublishAction}
\alias{ApplicationPublishAction}
\title{ApplicationPublishAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationPublishAction
}
\description{
Publishes an app into the public space for anyone to use.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
