% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationSummary.r
\docType{data}
\name{ApplicationSummary}
\alias{ApplicationSummary}
\title{ApplicationSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationSummary
}
\description{
An abbreviated Application object returned in request in application collection queries
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique id of this app. Comprised of the app name-version.}

\item{\code{executionSystem}}{The system id of the execution system where this app should run.}

\item{\code{lastModified}}{The date this application was last modified in ISO 8601 format.}

\item{\code{name}}{The name of the application. The name does not have to be unique, but the combination of name and version does.}

\item{\code{isPublic}}{Whether the application is public or private.}

\item{\code{revision}}{The number of times this application has been revised.}

\item{\code{shortDescription}}{The short description of this application.}

\item{\code{version}}{The version of the application in #.#.# format. While the version does not need to be unique, the combination of name and version does have to be unique.}
}}

\keyword{datasets}
