% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationUnpublishAction.r
\docType{data}
\name{ApplicationUnpublishAction}
\alias{ApplicationUnpublishAction}
\title{ApplicationUnpublishAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationUnpublishAction
}
\description{
Removes an app from the public scope.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
