% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppsApi.r
\docType{data}
\name{AppsApi}
\alias{AppsApi}
\title{Agave Apps API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
apps <- AppsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
AppsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addApp} Register and update new applications.

\strong{Usage}\preformatted{

`addAppPermission` Grant a user permission for an application.

**Usage**

}

\code{clearAppPermissions} Deletes all permissions on an application.

\strong{Usage}\preformatted{

`deleteApp` Deletes an application.

**Usage**

}

\code{deleteAppPermission} Deletes all permissions for the given user on an application.

\strong{Usage}\preformatted{

`getAppDetails` Get details of an application by its unique id.

**Usage**

}

\code{getAppPermission} Get a specific user permission for an application.

\strong{Usage}\preformatted{

`getAppSubmissionForm` Get a submission form for the named application.

**Usage**

}

\code{invokeAppAction} Edit an application.

\strong{Usage}\preformatted{

`listAppHistory` List the event history of this app

**Usage**

}

\code{listAppPermissions} Get the permission for this application.

\strong{Usage}\preformatted{

`listApps` List apps

**Usage**

}

\code{updateApp} Update an application.

\strong{Usage}\preformatted{

`updateAppPermission` Add or update a user permission for an application.

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::Application][rAgave::ApplicationAction][rAgave::ApplicationSummary][rAgave::HistoryEvent][rAgave::Permission]
}
\keyword{datasets}
