% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationType.r
\docType{data}
\name{AuthenticationType}
\alias{AuthenticationType}
\title{AuthenticationType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthenticationType
}
\description{
AuthenticationType Class
}
\keyword{datasets}
