% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchQueue.r
\docType{data}
\name{BatchQueue}
\alias{BatchQueue}
\title{BatchQueue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BatchQueue
}
\description{
BatchQueue Class
}
\section{Fields}{

\describe{
\item{\code{customDirectives}}{Any custom directives that should be appended to scheduler directives. ex. #$ -A TG-12345}

\item{\code{isDefault}}{Is this the default queue for the system.}

\item{\code{maxJobs}}{The maximum number of jobs that can be in queue at once.}

\item{\code{maxUserJobs}}{The maximum number of jobs per user that can be in queue at once.}

\item{\code{maxNodes}}{The max nodes available per node to jobs submitted to this queue.}

\item{\code{maxMemoryPerNode}}{The max memory available per node to jobs submitted to this queue.}

\item{\code{maxProcessorsPerNode}}{The max processors per node available to jobs submitted to this queue.}

\item{\code{maxRequestedTime}}{The max length of jobs submitted to this queue in hhh:mm:ss format.}

\item{\code{name}}{The name of the batch queue.}
}}

\keyword{datasets}
