% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchSchedulerType.r
\docType{data}
\name{BatchSchedulerType}
\alias{BatchSchedulerType}
\title{BatchSchedulerType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BatchSchedulerType
}
\description{
BatchSchedulerType Class
}
\keyword{datasets}
