% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.r
\docType{data}
\name{Client}
\alias{Client}
\title{Client Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Client
}
\description{
Client Class
}
\section{Fields}{

\describe{
\item{\code{key}}{The API key for this client.}

\item{\code{secret}}{The API secret for this client.}

\item{\code{callbackUrl}}{Callback URL for OAuth authentication grant.}

\item{\code{description}}{Description of the client. This will be shown to users when authentication via OAuth web flows}

\item{\code{name}}{The name of the client.}

\item{\code{tier}}{The access tier for this client. Usually "UNLIMITED"}
}}

\keyword{datasets}
