% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClientAPISubscription.r
\docType{data}
\name{ClientAPISubscription}
\alias{ClientAPISubscription}
\title{ClientAPISubscription Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ClientAPISubscription
}
\description{
ClientAPISubscription Class
}
\section{Fields}{

\describe{
\item{\code{apiContext}}{The base url path of the API.}

\item{\code{apiName}}{The name of the API.}

\item{\code{apiProvider}}{The user who registered the API.}

\item{\code{apiVersion}}{The current major version of the API. This is appended to the api_context to create the base API url.}

\item{\code{status}}{}

\item{\code{tier}}{}
}}

\keyword{datasets}
