% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClientAPISubscriptionStatusType.r
\docType{data}
\name{ClientAPISubscriptionStatusType}
\alias{ClientAPISubscriptionStatusType}
\title{ClientAPISubscriptionStatusType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ClientAPISubscriptionStatusType
}
\description{
ClientAPISubscriptionStatusType Class
}
\keyword{datasets}
