% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClientRequest.r
\docType{data}
\name{ClientRequest}
\alias{ClientRequest}
\title{ClientRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ClientRequest
}
\description{
ClientRequest Class
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the client.}

\item{\code{description}}{Description of the client. This will be shown to users when authentication via OAuth web flows}

\item{\code{tier}}{}

\item{\code{callbackUrl}}{Callback URL for OAuth authentication grant.}
}}

\keyword{datasets}
