% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClientsApi.r
\docType{data}
\name{ClientsApi}
\alias{ClientsApi}
\title{Clients operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ClientsApi
}
\description{
rAgave.Clients
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_client

add_client_api_subscription

clear_client_api_subscriptions

delete_client

get_client

list_client_api_subscriptions

list_clients

}
}

\keyword{datasets}
