% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClientSubscriptionRequest.r
\docType{data}
\name{ClientSubscriptionRequest}
\alias{ClientSubscriptionRequest}
\title{ClientSubscriptionRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ClientSubscriptionRequest
}
\description{
ClientSubscriptionRequest Class
}
\section{Fields}{

\describe{
\item{\code{apiName}}{The name of the API.}

\item{\code{apiProvider}}{The user who registered the API.}

\item{\code{apiVersion}}{The current major version of the API. This is appended to the api_context to create the base API url.}

\item{\code{tier}}{}
}}

\keyword{datasets}
