% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyApplicationResponse.r
\docType{data}
\name{EmptyApplicationResponse}
\alias{EmptyApplicationResponse}
\title{EmptyApplicationResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EmptyApplicationResponse
}
\description{
EmptyApplicationResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{empty result}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
