% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyClientResponse.r
\docType{data}
\name{EmptyClientResponse}
\alias{EmptyClientResponse}
\title{EmptyClientResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EmptyClientResponse
}
\description{
EmptyClientResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
