% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyRemoteFileResponse.r
\docType{data}
\name{EmptyRemoteFileResponse}
\alias{EmptyRemoteFileResponse}
\title{EmptyRemoteFileResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EmptyRemoteFileResponse
}
\description{
EmptyRemoteFileResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{empty result}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
