% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptySubscriptionResponse.r
\docType{data}
\name{EmptySubscriptionResponse}
\alias{EmptySubscriptionResponse}
\title{EmptySubscriptionResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EmptySubscriptionResponse
}
\description{
EmptySubscriptionResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
