% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptySystemResponse.r
\docType{data}
\name{EmptySystemResponse}
\alias{EmptySystemResponse}
\title{EmptySystemResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EmptySystemResponse
}
\description{
EmptySystemResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{empty result}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
