% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileAction.r
\docType{data}
\name{FileAction}
\alias{FileAction}
\title{FileAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileAction
}
\description{
A generic action to take on a file item
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{path}}{Name of new directory or target file or folder.}
}}

\keyword{datasets}
