% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileCopyAction.r
\docType{data}
\name{FileCopyAction}
\alias{FileCopyAction}
\title{FileCopyAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileCopyAction
}
\description{
Request for a file/folder to be copied on the target system. Metadata will be appended to the destination if it already exists. Otherwise, a new history will be created.
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{path}}{Name of new directory or target file or folder.}
}}

\keyword{datasets}
