% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHistoryResponse.r
\docType{data}
\name{FileHistoryResponse}
\alias{FileHistoryResponse}
\title{FileHistoryResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileHistoryResponse
}
\description{
FileHistoryResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{status}}{success or failure}

\item{\code{version}}{The current api version}
}}

\keyword{datasets}
