% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileImportRequest.r
\docType{data}
\name{FileImportRequest}
\alias{FileImportRequest}
\title{FileImportRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileImportRequest
}
\description{
File import request to copy a file from an accessible URL.
}
\section{Fields}{

\describe{
\item{\code{fileType}}{The file format this file is in. Defaults to raw. This will be used in file transform operations.}

\item{\code{notifications}}{Notifications to set with the successful acceptance of this upload request}

\item{\code{fileName}}{The name to give the upload file. If unset, takes the original file name.}

\item{\code{append}}{Boolean flag indicating this file should be appended to the target file}

\item{\code{urlToImport}}{The URL to import the file from. This parameter is used if not file is uploaded with this post.}
}}

\keyword{datasets}
