% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInfo.r
\docType{data}
\name{FileInfo}
\alias{FileInfo}
\title{FileInfo Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileInfo
}
\description{
FileInfo Class
}
\section{Fields}{

\describe{
\item{\code{format}}{}

\item{\code{lastModified}}{}

\item{\code{length}}{}

\item{\code{mimeType}}{}

\item{\code{name}}{}

\item{\code{path}}{}

\item{\code{permissions}}{}

\item{\code{system}}{}

\item{\code{type}}{}
}}

\keyword{datasets}
