% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileManagementActionType.r
\docType{data}
\name{FileManagementActionType}
\alias{FileManagementActionType}
\title{FileManagementActionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileManagementActionType
}
\description{
FileManagementActionType Class
}
\keyword{datasets}
