% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileMkdirAction.r
\docType{data}
\name{FileMkdirAction}
\alias{FileMkdirAction}
\title{FileMkdirAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileMkdirAction
}
\description{
Request for a new directory to be created on the target system. This will recursively create missing directories.
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{path}}{Name of new directory or target file or folder.}
}}

\keyword{datasets}
