% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileMoveAction.r
\docType{data}
\name{FileMoveAction}
\alias{FileMoveAction}
\title{FileMoveAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileMoveAction
}
\description{
Request for a file/folder to be moved on the target system. Metadata will be preserved after move.
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{path}}{Name of new directory or target file or folder.}
}}

\keyword{datasets}
