% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePermissionRequest.r
\docType{data}
\name{FilePermissionRequest}
\alias{FilePermissionRequest}
\title{FilePermissionRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FilePermissionRequest
}
\description{
The permission to apply for the given user onthe given resource
}
\section{Fields}{

\describe{
\item{\code{username}}{The name of the user for this permission object.}

\item{\code{permission}}{}
}}

\keyword{datasets}
