% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileRenameAction.r
\docType{data}
\name{FileRenameAction}
\alias{FileRenameAction}
\title{FileRenameAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileRenameAction
}
\description{
Request for a file/folder to be renamed on the target system. Metadata will be preserved after rename. Rename operations are only applied relative to the file/folder given in the URL. To rename a file/folder to a different path, see the FileMoveAction.
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{path}}{Name of new directory or target file or folder.}
}}

\keyword{datasets}
