% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesApi.r
\docType{data}
\name{FilesApi}
\alias{FilesApi}
\title{Files operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FilesApi
}
\description{
rAgave.Files
}
\section{Methods}{

\describe{

clear_file_item_permissions

delete

download

transfer

upload

invoke_file_action_on_default_system

invoke_file_item_action

list_file_item_history

list_file_item_history_on_default_system

list_file_item_permissions

list_file_item_permissions_on_default_system

list_file_items

update_file_item_permission

update_file_item_permissions_on_default_system

}
}

\keyword{datasets}
