% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesStatusType.r
\docType{data}
\name{FilesStatusType}
\alias{FilesStatusType}
\title{FilesStatusType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FilesStatusType
}
\description{
FilesStatusType Class
}
\keyword{datasets}
