% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileType.r
\docType{data}
\name{FileType}
\alias{FileType}
\title{FileType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileType
}
\description{
The type of the remote file/folder
}
\keyword{datasets}
