% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoryEvent.r
\docType{data}
\name{HistoryEvent}
\alias{HistoryEvent}
\title{HistoryEvent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HistoryEvent
}
\description{
A generic history event representing a change on a resource or one of its associations.
}
\section{Fields}{

\describe{
\item{\code{created}}{The date of the event.}

\item{\code{description}}{A brief description of the event details.}

\item{\code{status}}{The status of the file/folder after this event.}

\item{\code{progress}}{}
}}

\keyword{datasets}
