% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.r
\docType{data}
\name{Job}
\alias{Job}
\title{Job Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Job
}
\description{
Job Class
}
\section{Fields}{

\describe{
\item{\code{appId}}{The unique name of the application being run by this job. This must be a valid application that the calling user has permission to run.}

\item{\code{archive}}{Whether the output from this job should be archived. If true, all new files created by this application's execution will be archived to the archivePath in the user's default storage system.}

\item{\code{archivePath}}{The path of the archive folder for this job on default storage sytem of the user}

\item{\code{archiveSystem}}{The unique id of the storage system on which this job output will be staged.}

\item{\code{batchQueue}}{The queue to which this job should be submitted. This is optional and only applies when the execution system has a batch scheduler.}

\item{\code{endTime}}{The date the job stopped running due to termination, completion, or error in ISO 8601 format.}

\item{\code{executionSystem}}{The system id of the execution system.}

\item{\code{id}}{The unique id of the job.}

\item{\code{inputs}}{}

\item{\code{localId}}{The process or local job id of the job on the remote execution system.}

\item{\code{memoryPerNode}}{The requested memory for this application to run given in GB.}

\item{\code{message}}{The error message incurred when the job failed.}

\item{\code{name}}{The name of the job.}

\item{\code{nodeCount}}{The number of processors this application should utilize while running. If the application is not of executionType PARALLEL, this should be 1.}

\item{\code{notifications}}{An array of notifications you wish to receive.}

\item{\code{outputPath}}{Relative path of the job output data.}

\item{\code{owner}}{The job owner.}

\item{\code{parameters}}{}

\item{\code{processorsPerNode}}{The number of processors this application should utilize while running. If the application is not of executionType PARALLEL, this should be 1.}

\item{\code{maxRunTime}}{The requested compute time needed for this application to complete given in HH:mm:ss format.}

\item{\code{retries}}{The number of retires it took to submit this job.}

\item{\code{startTime}}{The date the job started in ISO 8601 format.}

\item{\code{status}}{The status of the job. Possible values are: PENDING, STAGING_INPUTS, CLEANING_UP, ARCHIVING, STAGING_JOB, FINISHED, KILLED, FAILED, STOPPED, RUNNING, PAUSED, QUEUED, SUBMITTING, STAGED, PROCESSING_INPUTS, ARCHIVING_FINISHED, ARCHIVING_FAILED}

\item{\code{submitTime}}{The date the job was submitted in ISO 8601 format.}

\item{\code{workPath}}{The directory on the remote execution system from which the job is running.}
}}

\keyword{datasets}
