% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobAction.r
\docType{data}
\name{JobAction}
\alias{JobAction}
\title{JobAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobAction
}
\description{
Requests that a management action take place against a job.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
