% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobActionType.r
\docType{data}
\name{JobActionType}
\alias{JobActionType}
\title{JobActionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobActionType
}
\description{
The type of action to perform on a job.
}
\keyword{datasets}
