% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobHistory.r
\docType{data}
\name{JobHistory}
\alias{JobHistory}
\title{JobHistory Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobHistory
}
\description{
JobHistory Class
}
\section{Fields}{

\describe{
\item{\code{created}}{The date of the event.}

\item{\code{description}}{A brief description of the event details.}

\item{\code{status}}{The status of the job after this event.}
}}

\keyword{datasets}
