% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobInputs.r
\docType{data}
\name{JobInputs}
\alias{JobInputs}
\title{JobInputs Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobInputs
}
\description{
JobInputs Class
}
\section{Fields}{

\describe{
\item{\code{parameter1}}{One or more inputs identified in the description of the app you're running.}
}}

\keyword{datasets}
