% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobParameters.r
\docType{data}
\name{JobParameters}
\alias{JobParameters}
\title{JobParameters Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobParameters
}
\description{
JobParameters Class
}
\section{Fields}{

\describe{
\item{\code{parameter1}}{One or more parameters identified in the description of the app you're running.}
}}

\keyword{datasets}
