% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResubmitAction.r
\docType{data}
\name{JobResubmitAction}
\alias{JobResubmitAction}
\title{JobResubmitAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobResubmitAction
}
\description{
Resubmits the given job as a new job with a unique id. The archive system and path are reset. All other fields remain unchanged. Notifications set for the original job will not carry over.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
