% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobsApi.r
\docType{data}
\name{JobsApi}
\alias{JobsApi}
\title{Agave Jobs API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
jobs <- JobsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
JobsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addJobPermission} Add a user permission for the given job.

\strong{Usage}\preformatted{

`clearJobPermissions` Deletes all permissions on an job.

**Usage**

}

\code{deleteJob} Deletes a job from history of the user

\strong{Usage}\preformatted{

`deleteJobPermission` Deletes all permissions for the given user on an job.

**Usage**

}

\code{downloadJobOutput} Download an output file from a specific job.

\strong{Usage}\preformatted{

`getJobDetails` Get details of the job with the specific job id.

**Usage**

}

\code{getJobHistory} Get the history of this job.

\strong{Usage}\preformatted{

`getJobPermission` Get a specific user permission for a job.

**Usage**

}

\code{getJobStatus} Get the status of the job.

\strong{Usage}\preformatted{

`listJobOutputs` List the output folder for a job.

**Usage**

}

\code{listJobPermissions} Get the permissions granted for this job.

\strong{Usage}\preformatted{

`listJobs` Get a list of jobs the authenticated user had submitted.

**Usage**

}

\code{resubmitJob} Resubmits the job with the given job id as a new job with a unique id.

\strong{Usage}\preformatted{

`submitJob` Submit a new job request.

**Usage**

}

\code{updateJobPermission} Update a user permission for the given job and username

\strong{Usage}\preformatted{
}
}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::FileInfo][rAgave::Job][rAgave::JobHistory][rAgave::JobRequest][rAgave::JobResubmitAction][rAgave::JobStatusSummary][rAgave::Permission]
}
\keyword{datasets}
