% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobStatusSummary.r
\docType{data}
\name{JobStatusSummary}
\alias{JobStatusSummary}
\title{JobStatusSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobStatusSummary
}
\description{
JobStatusSummary Class
}
\section{Fields}{

\describe{
\item{\code{id}}{The unique id of the job.}

\item{\code{status}}{The status of the job. Possible values are: PENDING, STAGING_INPUTS, CLEANING_UP, ARCHIVING, STAGING_JOB, FINISHED, KILLED, FAILED, STOPPED, RUNNING, PAUSED, QUEUED, SUBMITTING, STAGED, PROCESSING_INPUTS, ARCHIVING_FINISHED, ARCHIVING_FAILED}
}}

\keyword{datasets}
