% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobStatusType.r
\docType{data}
\name{JobStatusType}
\alias{JobStatusType}
\title{JobStatusType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobStatusType
}
\description{
JobStatusType Class
}
\keyword{datasets}
