% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobStopAction.r
\docType{data}
\name{JobStopAction}
\alias{JobStopAction}
\title{JobStopAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobStopAction
}
\description{
Requests that a currently active job be killed.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
