% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobSummary.r
\docType{data}
\name{JobSummary}
\alias{JobSummary}
\title{JobSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobSummary
}
\description{
JobSummary Class
}
\section{Fields}{

\describe{
\item{\code{appId}}{The unique name of the application being run by this job. This must be a valid application that the calling user has permission to run.}

\item{\code{endTime}}{The date the job ended in ISO 8601 format.}

\item{\code{executionSystem}}{The system id of the execution system.}

\item{\code{id}}{The unique id of the job.}

\item{\code{name}}{The name of the job.}

\item{\code{owner}}{The job owner.}

\item{\code{startTime}}{The date the job started in ISO 8601 format.}

\item{\code{status}}{The status of the job. Possible values are: PENDING, STAGING_INPUTS, CLEANING_UP, ARCHIVING, STAGING_JOB, FINISHED, KILLED, FAILED, STOPPED, RUNNING, PAUSED, QUEUED, SUBMITTING, STAGED, PROCESSING_INPUTS, ARCHIVING_FINISHED, ARCHIVING_FAILED}
}}

\keyword{datasets}
