% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaApi.r
\docType{data}
\name{MetaApi}
\alias{MetaApi}
\title{Agave Meta API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
meta <- MetaApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
MetaApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addMetadata} Update or Add new Metadata.

\strong{Usage}\preformatted{

`addMetadataPermission` Add a user permission for the given metadata.

**Usage**

}

\code{addMetadataSchema} Add a new Metadata Schema.

\strong{Usage}\preformatted{

`addMetadataSchemaPermission` Add a user permission for the given schema.

**Usage**

}

\code{clearMetadataPermissions} Deletes all permissions on the given metadata.

\strong{Usage}\preformatted{

`clearMetadataSchemaPermissions` Deletes all permissions on the given schema.

**Usage**

}

\code{deleteMetadata} Remove Metadata from the system.

\strong{Usage}\preformatted{

`deleteMetadataPermission` Removes user permissions for a user on a given metadata resource.

**Usage**

}

\code{deleteMetadataSchema} Remove Metadata Schema from the system.

\strong{Usage}\preformatted{

`deleteSchemaPermission` Deletes all metadata schema permissions on the given metadata.

**Usage**

}

\code{getMetadata} Retrieve Metadata.

\strong{Usage}\preformatted{

`getMetadataPermission` Get the user permission for this metadata.

**Usage**

}

\code{getMetadataSchema} Retrieve Metadata Schemata.

\strong{Usage}\preformatted{

`getMetadataSchemaPermission` Get the user permission for this schema.

**Usage**

}

\code{listMetadata} List and/or search metadata.

\strong{Usage}\preformatted{

`listMetadataPermissions` Get the permission ACL for this metadata.

**Usage**

}

\code{listMetadataSchema} List and query Metadata Schemata.

\strong{Usage}\preformatted{

`listMetadataSchemaPermissions` Get the permission for this schema.

**Usage**

}

\code{updateMetadata} Update or Add new Metadata.

\strong{Usage}\preformatted{

`updateMetadataPermission` Update a user permission for the given metadata.

**Usage**

}

\code{updateMetadataSchema} Update an existing Metadata Schema.

\strong{Usage}\preformatted{

`updateMetadataSchemaPermission` Add or update a user permission for the given metadata schema.

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::Metadata][rAgave::MetadataSchema][rAgave::Permission]
}
\keyword{datasets}
