% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.r
\docType{data}
\name{Metadata}
\alias{Metadata}
\title{Metadata Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Metadata
}
\description{
Metadata Class
}
\section{Fields}{

\describe{
\item{\code{associationIds}}{UUIDs of associated Agave entities, including the Data to which this Metadata belongs.}

\item{\code{created}}{A timestamp indicating when this Metadata was created in the metadata store.}

\item{\code{internalUsername}}{The name of the Internal User, if any, who owns this metadata.}

\item{\code{lastUpdated}}{A timestamp indicating when this Metadata was last updated in the metadata store.}

\item{\code{name}}{The name of this metadata}

\item{\code{owner}}{The API user who owns this Metadata.}

\item{\code{uuid}}{The UUID for this Metadata.}

\item{\code{value}}{A free text or JSON string containing the metadata stored for the given associationIds}
}}

\keyword{datasets}
