% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataSchema.r
\docType{data}
\name{MetadataSchema}
\alias{MetadataSchema}
\title{MetadataSchema Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetadataSchema
}
\description{
MetadataSchema Class
}
\section{Fields}{

\describe{
\item{\code{created}}{A timestamp indicating when this Metadata was created in the metadata schema store.}

\item{\code{internalUsername}}{The name of the Internal User, if any, who owns this schema.}

\item{\code{lastUpdated}}{A timestamp indicating when this Metadata was last updated in the metadata schema store.}

\item{\code{owner}}{The API user who owns this Schema.}

\item{\code{schema}}{A JSON Schema}

\item{\code{uuid}}{The UUID for this Schema.}
}}

\keyword{datasets}
