% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitoringTaskCheck.r
\docType{data}
\name{MonitoringTaskCheck}
\alias{MonitoringTaskCheck}
\title{MonitoringTaskCheck Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MonitoringTaskCheck
}
\description{
An individual check made for a monitoring task.
}
\section{Fields}{

\describe{
\item{\code{created}}{A timestamp indicating when this monitor check was created.}

\item{\code{id}}{The UUID for this monitor check.}

\item{\code{message}}{The error message if this monitor check failed.}

\item{\code{result}}{}
}}

\keyword{datasets}
