% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitoringTaskDetails.r
\docType{data}
\name{MonitoringTaskDetails}
\alias{MonitoringTaskDetails}
\title{MonitoringTaskDetails Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MonitoringTaskDetails
}
\description{
Detailed description of a monitoring task
}
\section{Fields}{

\describe{
\item{\code{active}}{Whether this monitor is currently active.}

\item{\code{frequency}}{The interval in minutes on which this monitor will run. Minimum is 5. Default is 720.}

\item{\code{internalUsername}}{Internal user account used to perform the check.}

\item{\code{target}}{The id of the sytem to be monitored. This must be an active system registered with the Systems service.}

\item{\code{updateSystemStatus}}{Whether this Monitor should update the system status when the results change. You must have the ADMIN role on the target system to use this feature.}

\item{\code{id}}{The unique id of the monitor}
}}

\keyword{datasets}
