% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleApplicationResponse.r
\docType{data}
\name{MultipleApplicationResponse}
\alias{MultipleApplicationResponse}
\title{MultipleApplicationResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleApplicationResponse
}
\description{
MultipleApplicationResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{response body}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
