% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleClientResponse.r
\docType{data}
\name{MultipleClientResponse}
\alias{MultipleClientResponse}
\title{MultipleClientResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleClientResponse
}
\description{
MultipleClientResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{List of clients}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
