% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplePermissionResponse.r
\docType{data}
\name{MultiplePermissionResponse}
\alias{MultiplePermissionResponse}
\title{MultiplePermissionResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiplePermissionResponse
}
\description{
MultiplePermissionResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{One or more file permissions}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
