% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleRemoteFileResponse.r
\docType{data}
\name{MultipleRemoteFileResponse}
\alias{MultipleRemoteFileResponse}
\title{MultipleRemoteFileResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleRemoteFileResponse
}
\description{
MultipleRemoteFileResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{response body}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
