% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleSubscriptionResponse.r
\docType{data}
\name{MultipleSubscriptionResponse}
\alias{MultipleSubscriptionResponse}
\title{MultipleSubscriptionResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleSubscriptionResponse
}
\description{
MultipleSubscriptionResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{List of subscriptions}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
