% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleSystemResponse.r
\docType{data}
\name{MultipleSystemResponse}
\alias{MultipleSystemResponse}
\title{MultipleSystemResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleSystemResponse
}
\description{
MultipleSystemResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{response body}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
