% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleSystemRoleResponse.r
\docType{data}
\name{MultipleSystemRoleResponse}
\alias{MultipleSystemRoleResponse}
\title{MultipleSystemRoleResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleSystemRoleResponse
}
\description{
MultipleSystemRoleResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{All matching roles for this system}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
