% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTransformResponse.r
\docType{data}
\name{MultipleTransformResponse}
\alias{MultipleTransformResponse}
\title{MultipleTransformResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultipleTransformResponse
}
\description{
MultipleTransformResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{response body}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
