% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notification.r
\docType{data}
\name{Notification}
\alias{Notification}
\title{Notification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Notification
}
\description{
Notification Class
}
\section{Fields}{

\describe{
\item{\code{event}}{The event you wish to recieve a notification about. Specify * for all events.}

\item{\code{persistent}}{Whether this event should remain active after it triggers the first time.}

\item{\code{url}}{The URI to notify when the import is complete. This can be an email address or URL. If a URL is given, a POST will be made to this address. URL templating is supported. Valid template values are: ${JOB_STATUS}, ${JOB_URL}, ${JOB_ID}, ${JOB_SYSTEM}, ${JOB_NAME}, ${JOB_START_TIME}, ${JOB_END_TIME}, ${JOB_SUBMIT_TIME}, ${JOB_ARCHIVE_PATH}, ${JOB_ARCHIVE_URL}, ${JOB_ERROR}}
}}

\keyword{datasets}
