% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotificationRequest.r
\docType{data}
\name{NotificationRequest}
\alias{NotificationRequest}
\title{NotificationRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NotificationRequest
}
\description{
NotificationRequest Class
}
\section{Fields}{

\describe{
\item{\code{associatedUuid}}{UUID of resource to whome the event applies.}

\item{\code{url}}{The url or email address that will be notified of the event.}

\item{\code{persistent}}{Whether this notification should stay active after it fires the first time.}
}}

\keyword{datasets}
