% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotificationResponse.r
\docType{data}
\name{NotificationResponse}
\alias{NotificationResponse}
\title{NotificationResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NotificationResponse
}
\description{
NotificationResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}

\item{\code{version}}{API version number}
}}

\keyword{datasets}
