% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permission.r
\docType{data}
\name{Permission}
\alias{Permission}
\title{Permission Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Permission
}
\description{
A generic permission applied to any resource in the API
}
\section{Fields}{

\describe{
\item{\code{username}}{The username to whom this permission grant applies}

\item{\code{internalUsername}}{The internal user associated with this permission}

\item{\code{permission}}{}

\item{\code{created}}{The date the permission was created}

\item{\code{lastUpdated}}{The date the permission was last updated}

\item{\code{id}}{The unique id of this permission}
}}

\keyword{datasets}
