% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermissionRequest.r
\docType{data}
\name{PermissionRequest}
\alias{PermissionRequest}
\title{PermissionRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PermissionRequest
}
\description{
A permission request for a given user on the given resource
}
\section{Fields}{

\describe{
\item{\code{username}}{The name of the user for this permission object.}

\item{\code{permission}}{}
}}

\keyword{datasets}
