% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermissionStanza.r
\docType{data}
\name{PermissionStanza}
\alias{PermissionStanza}
\title{PermissionStanza Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PermissionStanza
}
\description{
A minimal permission stanza defining read and write permissions
}
\section{Fields}{

\describe{
\item{\code{read}}{Is the resource readable}

\item{\code{write}}{Is the resource writable}
}}

\keyword{datasets}
