% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermissionType.r
\docType{data}
\name{PermissionType}
\alias{PermissionType}
\title{PermissionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PermissionType
}
\description{
PermissionType Class
}
\keyword{datasets}
