% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostIt.r
\docType{data}
\name{PostIt}
\alias{PostIt}
\title{PostIt Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PostIt
}
\description{
PostIt Class
}
\section{Fields}{

\describe{
\item{\code{created}}{The creation date in ISO 8601 format.}

\item{\code{creator}}{The api user who made the PostIt creation request.}

\item{\code{expires}}{The expiration date in ISO 8601 format.}

\item{\code{internalUsername}}{The username of the internal user attached to this PostIt.}

\item{\code{method}}{The HTTP method to be invoked on the target URL when this PostIt is redeemed}

\item{\code{postit}}{The PostIt nonce.}

\item{\code{remainingUses}}{The number of invocations remaining on this token.}

\item{\code{url}}{The url that will be invoked when the PostIt is redeemed.}

\item{\code{noauth}}{If true, the provided url will be called without authentication. Default is false}
}}

\keyword{datasets}
