% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostItRequest.r
\docType{data}
\name{PostItRequest}
\alias{PostItRequest}
\title{PostItRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PostItRequest
}
\description{
PostItRequest Class
}
\section{Fields}{

\describe{
\item{\code{url}}{The url that will be invoked when the PostIt is redeemed.}

\item{\code{internalUsername}}{The username of the internal user attached to this PostIt.}

\item{\code{method}}{}

\item{\code{lifetime}}{The maximum lifetime in seconds of this PostIt on this token. Defaults to 2592000 (30 days)}

\item{\code{maxUses}}{The maximum number of invocations remaining on this PostIt. Defaults to no limit}

\item{\code{noauth}}{If true, the provided url will be called without authentication. Default is false}
}}

\keyword{datasets}
