% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.r
\docType{data}
\name{Profile}
\alias{Profile}
\title{Profile Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Profile
}
\description{
Profile Class
}
\section{Fields}{

\describe{
\item{\code{uuid}}{}
}}

\keyword{datasets}
