% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilesApi.r
\docType{data}
\name{ProfilesApi}
\alias{ProfilesApi}
\title{Agave Profiles API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
profiles <- ProfilesApi$new(apiClient, cache, responseType="list")
}
\description{
ProfilesApi Class

The Profiles API is a user directory service for the Agave Platform.
}
\details{
The Profiles API is a user directory service for the Agave Platform. Users
may query it for their own profile and search for the profile(s) of other
users. Depending on the configuration of your tenant, additional features
such as profile editing, profile creation, and password reset may be enabled
as well.
}
\section{Methods}{

\describe{

\code{$addProfile(body)} Create a new user profile. This may be disabled depending
on the configuration of your specific tenant.

\strong{Usage}\preformatted{profile <- Profile$new(username="username",
                       password="password",
                       email="email@email.com")
api$profiles$addProfile(body=profile)
}

\code{$deleteProfile(username=username)} deletes a user profile. The user's data and history is
maintained. Depending on the backend identity provider, the user account
may or may not be avaialble for reuse/reactivation.

\strong{Usage}\preformatted{api$profiles$deleteProfile(username="username")
}

\code{$getProfile(username)} get the full profile of a given user.

\strong{Usage}\preformatted{api$profiles$getProfile()
api$profiles$getProfile(username="username")
}

\code{$listProfiles(username, name, email)} list user profiles, filtering by
partial match on username, full name, or email address.

\strong{Usage}\preformatted{api$profiles$listProfiles()
api$profiles$listProfiles(username="test")
api$profiles$listProfiles(name="Test")
api$profiles$listProfiles(email="test@")
}

\code{$updateProfile} Update an existing user profile. This may be disabled
depending on the configuration of your specific tenant.

\strong{Usage}\preformatted{profile <- Profile$new(username="username",
                       password="newpassword",
                       email="email@email.com")
api$profiles$updateProfile(username="username", body=profile)
}

}
}

\seealso{
\link[rAgave:Profile]{rAgave::Profile} \link[rAgave:ClientsApi]{rAgave::ClientsApi} \link[rAgave:Client]{rAgave::Client} \link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave}
}
\keyword{datasets}
