% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleClientResponse.r
\docType{data}
\name{SingleClientResponse}
\alias{SingleClientResponse}
\title{SingleClientResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleClientResponse
}
\description{
SingleClientResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
