% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleInternalUserResponse.r
\docType{data}
\name{SingleInternalUserResponse}
\alias{SingleInternalUserResponse}
\title{SingleInternalUserResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleInternalUserResponse
}
\description{
SingleInternalUserResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
