% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePostItResponse.r
\docType{data}
\name{SinglePostItResponse}
\alias{SinglePostItResponse}
\title{SinglePostItResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SinglePostItResponse
}
\description{
SinglePostItResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
