% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleProfileResponse.r
\docType{data}
\name{SingleProfileResponse}
\alias{SingleProfileResponse}
\title{SingleProfileResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleProfileResponse
}
\description{
SingleProfileResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
