% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRemoteFileResponse.r
\docType{data}
\name{SingleRemoteFileResponse}
\alias{SingleRemoteFileResponse}
\title{SingleRemoteFileResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleRemoteFileResponse
}
\description{
SingleRemoteFileResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
