% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSubscriptionResponse.r
\docType{data}
\name{SingleSubscriptionResponse}
\alias{SingleSubscriptionResponse}
\title{SingleSubscriptionResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleSubscriptionResponse
}
\description{
SingleSubscriptionResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
