% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSystemResponse.r
\docType{data}
\name{SingleSystemResponse}
\alias{SingleSystemResponse}
\title{SingleSystemResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleSystemResponse
}
\description{
SingleSystemResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
