% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTransformResponse.r
\docType{data}
\name{SingleTransformResponse}
\alias{SingleTransformResponse}
\title{SingleTransformResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SingleTransformResponse
}
\description{
SingleTransformResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
