% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemAction.r
\docType{data}
\name{SystemAction}
\alias{SystemAction}
\title{SystemAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemAction
}
\description{
A generic action to apply to a system
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
