% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemActionType.r
\docType{data}
\name{SystemActionType}
\alias{SystemActionType}
\title{SystemActionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemActionType
}
\description{
The actions available on a system.
}
\keyword{datasets}
