% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemAuthConfigServer.r
\docType{data}
\name{SystemAuthConfigServer}
\alias{SystemAuthConfigServer}
\title{SystemAuthConfigServer Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemAuthConfigServer
}
\description{
SystemAuthConfigServer Class
}
\section{Fields}{

\describe{
\item{\code{endpoint}}{The hostname or ip address of the remote server.}

\item{\code{port}}{The port number of the remote server..}

\item{\code{protocol}}{}
}}

\keyword{datasets}
