% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemAuthConfigServerProtocolType.r
\docType{data}
\name{SystemAuthConfigServerProtocolType}
\alias{SystemAuthConfigServerProtocolType}
\title{SystemAuthConfigServerProtocolType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemAuthConfigServerProtocolType
}
\description{
SystemAuthConfigServerProtocolType Class
}
\keyword{datasets}
