% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCloneAction.r
\docType{data}
\name{SystemCloneAction}
\alias{SystemCloneAction}
\title{SystemCloneAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemCloneAction
}
\description{
Clone an existing system as a new system. Auth credentials will not be cloned on the new system.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
