% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCredential.r
\docType{data}
\name{SystemCredential}
\alias{SystemCredential}
\title{SystemCredential Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemCredential
}
\description{
Authentication credential for the default or internal user on a system
}
\section{Fields}{

\describe{
\item{\code{credential}}{The credential used to authenticate to the remote system. Depending on the authentication protocol of the remote system, this could be an OAuth Token, X.509 certificate, Kerberose token, or an private key..}

\item{\code{expirationDate}}{The date the credential expires in ISO 8601 format.}

\item{\code{internalUsername}}{The username of the internal user associated with this credential.}

\item{\code{isDefault}}{Is this the default credential for this internal user of this type on this system?}

\item{\code{parentType}}{}

\item{\code{password}}{The password on the remote system used to authenticate.}

\item{\code{publicKey}}{The public ssh key used to authenticate to the remote system.}

\item{\code{privateKey}}{The public ssh key used to authenticate to the remote system..}

\item{\code{server}}{}

\item{\code{type}}{}

\item{\code{username}}{The local username on the remote system used to authenticate.}

\item{\code{valid}}{Is the credential still valid or has it expired?.}

\item{\code{caCerts}}{A public URL to the location of the trusted CA certs to use with this credential.}
}}

\keyword{datasets}
