% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCredentialsResponse.r
\docType{data}
\name{SystemCredentialsResponse}
\alias{SystemCredentialsResponse}
\title{SystemCredentialsResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemCredentialsResponse
}
\description{
SystemCredentialsResponse Class
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}
}}

\keyword{datasets}
