% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemExecutionType.r
\docType{data}
\name{SystemExecutionType}
\alias{SystemExecutionType}
\title{SystemExecutionType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemExecutionType
}
\description{
SystemExecutionType Class
}
\keyword{datasets}
