% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemLoginAuthenticationType.r
\docType{data}
\name{SystemLoginAuthenticationType}
\alias{SystemLoginAuthenticationType}
\title{SystemLoginAuthenticationType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemLoginAuthenticationType
}
\description{
SystemLoginAuthenticationType Class
}
\keyword{datasets}
