% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemLoginConfig.r
\docType{data}
\name{SystemLoginConfig}
\alias{SystemLoginConfig}
\title{SystemLoginConfig Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemLoginConfig
}
\description{
SystemLoginConfig Class
}
\section{Fields}{

\describe{
\item{\code{auth}}{}

\item{\code{host}}{The hostname or ip address of the submission server}

\item{\code{port}}{The port number of the submission server.}

\item{\code{protocol}}{}
}}

\keyword{datasets}
