% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemLoginProtocolType.r
\docType{data}
\name{SystemLoginProtocolType}
\alias{SystemLoginProtocolType}
\title{SystemLoginProtocolType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemLoginProtocolType
}
\description{
SystemLoginProtocolType Class
}
\keyword{datasets}
