% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemPublishAction.r
\docType{data}
\name{SystemPublishAction}
\alias{SystemPublishAction}
\title{SystemPublishAction Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemPublishAction
}
\description{
Promote the system into the public space for all tenant users to utilize as a shared resource.
}
\section{Fields}{

\describe{
\item{\code{action}}{}
}}

\keyword{datasets}
