% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRequest.r
\docType{data}
\name{SystemRequest}
\alias{SystemRequest}
\title{SystemRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemRequest
}
\description{
SystemRequest Class
}
\section{Fields}{

\describe{
\item{\code{description}}{Verbose description of this system.}

\item{\code{environment}}{Environment variables to set upon login prior to job submission.}

\item{\code{executionType}}{}

\item{\code{id}}{Unique identifier for this system.}

\item{\code{login}}{}

\item{\code{maxSystemJobs}}{The maximum number of jobs that can be simultaneously run on the system across all queues.}

\item{\code{maxSystemJobsPerUser}}{The maximum number of jobs that can be simultaneously run on the system across all queues by a single user.}

\item{\code{name}}{Common name for this system.}

\item{\code{queues}}{The execution paradigm used to run jobs on this system.}

\item{\code{scheduler}}{}

\item{\code{scratchDir}}{The scratch directory where job execution directories will be created at runtime. The workDir is used if this is not specified.}

\item{\code{site}}{The site associated with this system.}

\item{\code{startupScript}}{Script to be run after login and prior to execution.}

\item{\code{status}}{}

\item{\code{storage}}{}

\item{\code{type}}{}

\item{\code{workDir}}{The work directory where job execution directories will be created at runtime. This is used if scratchDir is not specified. If neither are specified, the job directory will be created in the system homeDir.}
}}

\keyword{datasets}
