% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRole.r
\docType{data}
\name{SystemRole}
\alias{SystemRole}
\title{SystemRole Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemRole
}
\description{
A role granted a user on a system
}
\section{Fields}{

\describe{
\item{\code{username}}{The username to whom the role applies.}

\item{\code{role}}{}

\item{\code{created}}{The date the role was first granted}

\item{\code{lastUpdated}}{The date the role was last updated.}

\item{\code{id}}{The unique id of this role grant}
}}

\keyword{datasets}
