% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRoleRequest.r
\docType{data}
\name{SystemRoleRequest}
\alias{SystemRoleRequest}
\title{SystemRoleRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemRoleRequest
}
\description{
A basic role granted to a user for a particular system
}
\section{Fields}{

\describe{
\item{\code{username}}{The username of the api user granted this role.}

\item{\code{role}}{}
}}

\keyword{datasets}
