% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRoleType.r
\docType{data}
\name{SystemRoleType}
\alias{SystemRoleType}
\title{SystemRoleType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemRoleType
}
\description{
SystemRoleType Class
}
\keyword{datasets}
