% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemsApi.r
\docType{data}
\name{SystemsApi}
\alias{SystemsApi}
\title{Agave Systems API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
systems <- SystemsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
SystemsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addExecutionSystem} Add a new execution system

\strong{Usage}\preformatted{

`addStorageSystem` Add  a storage system

**Usage**

}

\code{addSystemCredential} Add an internal user's credential on a system. This applies both to storage and, if applicable, login credentials.

\strong{Usage}\preformatted{

`addSystemRole` Adds a role on the system for the given user

**Usage**

}

\code{clearSystemAuthCredentials} Deletes all credentials registered to a system.

\strong{Usage}\preformatted{

`clearSystemAuthCredentialsForInternalUser` Deletes all internal user credentials registered to a system.

**Usage**

}

\code{clearSystemRoles} Deletes all roles on a system.

\strong{Usage}\preformatted{

`deleteSystem` Delete a system.

**Usage**

}

\code{deleteSystemAuthCredentialForInternalUser} Deletes the internal user credentials for the given credential type on a system.

\strong{Usage}\preformatted{

`deleteSystemRole` Deletes all roles for a user on a system.

**Usage**

}

\code{getSystemCredential} Get the internal user credential of the given type on the system.

\strong{Usage}\preformatted{

`getSystemDetails` Find information about an individual system.

**Usage**

}

\code{getSystemRole} Get a specific user role on this system.

\strong{Usage}\preformatted{

`invokeSystemAction` Perform a management action on the system.

**Usage**

}

\code{listCredentialsForInternalUser} Get a list of all internal users and their credentials on this system.

\strong{Usage}\preformatted{

`listSystemCredentials` Get a list of all internal user credentials on this system.

**Usage**

}

\code{listSystemRoles} Get a list of all users and their roles on this system.

\strong{Usage}\preformatted{

`listSystems` Show all systems available to the user.

**Usage**

}

\code{updateSystem} Update a system description

\strong{Usage}\preformatted{

`updateSystemCredential` Update an internal user credential on a system.

**Usage**

}

\code{updateSystemCredentialOfType} Add or update a credential of the given type on a system.

\strong{Usage}\preformatted{

`updateSystemRole` Update a user role on a system.

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::ExecutionSystem][rAgave::MultipleSystemRoleResponse][rAgave::StorageSystem][rAgave::System][rAgave::SystemAction][rAgave::SystemAuthConfig][rAgave::SystemCredential][rAgave::SystemCredentialsResponse][rAgave::SystemRequest][rAgave::SystemRole][rAgave::SystemRoleRequest][rAgave::SystemSummary]
}
\keyword{datasets}
