% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemStatusType.r
\docType{data}
\name{SystemStatusType}
\alias{SystemStatusType}
\title{SystemStatusType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemStatusType
}
\description{
SystemStatusType Class
}
\keyword{datasets}
