% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemStorageConfig.r
\docType{data}
\name{SystemStorageConfig}
\alias{SystemStorageConfig}
\title{SystemStorageConfig Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemStorageConfig
}
\description{
SystemStorageConfig Class
}
\section{Fields}{

\describe{
\item{\code{auth}}{}

\item{\code{containerName}}{The name of the bucket to interact with when using an object store.}

\item{\code{homeDir}}{The path on the remote system to use as the home directory for all API requests. For cases where the home directory should be dynamically defined, the ${USERNAME} token can be included in the homeDir value to represent the username of the authenticated user. ex. /home/${USERNAME}.}

\item{\code{host}}{The hostname or ip address of the storage server}

\item{\code{port}}{The port number of the storage server.}

\item{\code{mirror}}{Whether the permissions set on the server should be pushed to the storage system itself. This only applies to IRODS and AWS systems.}

\item{\code{protocol}}{}

\item{\code{publicAppsDir}}{The path on the remote system where apps will be stored if this is set as the default public storage system.}

\item{\code{resource}}{The name of the target resource when defining an IRODS system.}

\item{\code{rootDir}}{The path on the remote system to use as the root for all API requests.}

\item{\code{zone}}{The name of the zone when defining an IRODS system.}
}}

\keyword{datasets}
