% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemStorageProtocolType.r
\docType{data}
\name{SystemStorageProtocolType}
\alias{SystemStorageProtocolType}
\title{SystemStorageProtocolType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemStorageProtocolType
}
\description{
SystemStorageProtocolType Class
}
\keyword{datasets}
