% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemSummary.r
\docType{data}
\name{SystemSummary}
\alias{SystemSummary}
\title{SystemSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemSummary
}
\description{
SystemSummary Class
}
\section{Fields}{

\describe{
\item{\code{description}}{Verbose description of this system.}

\item{\code{id}}{Unique identifier for this system.}

\item{\code{isDefault}}{Is the system the default for the authenticated user?}

\item{\code{isPublic}}{Is the system publicly available?}

\item{\code{name}}{Common name for this system.}

\item{\code{status}}{}

\item{\code{type}}{}
}}

\keyword{datasets}
