% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemType.r
\docType{data}
\name{SystemType}
\alias{SystemType}
\title{SystemType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SystemType
}
\description{
SystemType Class
}
\keyword{datasets}
