% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag.r
\docType{data}
\name{Tag}
\alias{Tag}
\title{Tag Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tag
}
\description{
Base tag object
}
\section{Fields}{

\describe{
\item{\code{id}}{uuid of this tag}

\item{\code{name}}{name of tag}

\item{\code{created}}{Date tag was created}

\item{\code{lastUpdated}}{Date tag was last updated}

\item{\code{resources}}{The list of resource uuid associated with this tag}
}}

\keyword{datasets}
