% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagApi.r
\docType{data}
\name{TagApi}
\alias{TagApi}
\title{Agave Tag API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tag <- TagApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
TagApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addTag} Creat a new tag

\strong{Usage}\preformatted{

`addTagResource` Add resources to tag

**Usage**

}

\code{clearTagPermissions} Clear tag permissions

\strong{Usage}\preformatted{

`clearTagResources` Clear all tagged resources

**Usage**

}

\code{deleteTag} Delete tag

\strong{Usage}\preformatted{

`deleteTagPermissionForUser` Remove user tag permission

**Usage**

}

\code{getTagDetails} Tag Details

\strong{Usage}\preformatted{

`getTagPermissionForUser` List tag permissions for a user

**Usage**

}

\code{getTaggedResource} Get tagged item details

\strong{Usage}\preformatted{

`listTagHistory` Tag history

**Usage**

}

\code{listTagPermissions} List permissionss set for a given tag id

\strong{Usage}\preformatted{

`listTagResources` Tagged items

**Usage**

}

\code{listTags} Available tags

\strong{Usage}\preformatted{

`removeTagResource` Remove resource from tag

**Usage**

}

\code{updateTag} Update tag

\strong{Usage}\preformatted{

`updateTagPermission` Grant permission

**Usage**

}

\code{updateTagPermissionForUser} Update user permission on tag

\strong{Usage}\preformatted{

`updateTagResource` Add resource to tag

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::Permission][rAgave::Tag][rAgave::TagHistory][rAgave::TagResource][rAgave::TagSummary]
}
\keyword{datasets}
