% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagHistory.r
\docType{data}
\name{TagHistory}
\alias{TagHistory}
\title{TagHistory Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TagHistory
}
\description{
Individual tag event
}
\section{Fields}{

\describe{
\item{\code{created}}{uuid of the referenced resource}

\item{\code{event}}{event that occurred on this tag}

\item{\code{description}}{description of the event which occurred}

\item{\code{id}}{unique identifier of the event}
}}

\keyword{datasets}
