% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagResource.r
\docType{data}
\name{TagResource}
\alias{TagResource}
\title{TagResource Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TagResource
}
\description{
Resource to which a tag has been associated
}
\section{Fields}{

\describe{
\item{\code{id}}{uuid of the API resource}
}}

\keyword{datasets}
