% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagSummary.r
\docType{data}
\name{TagSummary}
\alias{TagSummary}
\title{TagSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TagSummary
}
\description{
Tag object without resource list
}
\section{Fields}{

\describe{
\item{\code{name}}{name of tag}

\item{\code{created}}{Date tag was created}

\item{\code{lastUpdated}}{Date tag was last updated}
}}

\keyword{datasets}
