% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tenant.r
\docType{data}
\name{Tenant}
\alias{Tenant}
\title{Tenant Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tenant
}
\description{
Tenant Class
}
\section{Fields}{

\describe{
\item{\code{id}}{uuid of the tenant}

\item{\code{name}}{human readable name of the tenant organization}

\item{\code{baseUrl}}{The base url for all tenant requests. This should always be SSL enabled.}

\item{\code{code}}{The code by which this tenant is known, eg. "agave.prod"}

\item{\code{created}}{When the tenant was created in ISO8601 format}

\item{\code{lastUpdated}}{When the tenant was last updated in ISO8601 format}

\item{\code{contact}}{\link[rAgave:TenantContact]{rAgave::TenantContact} representing who to contact about tenant related issues}
}}

\seealso{
\link[rAgave:TenantContact]{rAgave::TenantContact}
}
\keyword{datasets}
