% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TenantContact.r
\docType{data}
\name{TenantContact}
\alias{TenantContact}
\title{TenantContact Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TenantContact
}
\description{
Tenant contact information
}
\section{Fields}{

\describe{
\item{\code{name}}{Full name of the contact}

\item{\code{email}}{The base url for all tenant requests. This should always be SSL enabled.}

\item{\code{url}}{The code by which this tenant is known, eg. agave.prod}

\item{\code{type}}{The type of contact this user represents.}

\item{\code{primary}}{Whether the contact is the primary contact for the tenant.}

\item{\code{contact}}{}
}}

\keyword{datasets}
