% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TenantsApi.r
\docType{data}
\name{TenantsApi}
\alias{TenantsApi}
\title{Tenants operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TenantsApi
}
\description{
TenantsApi Class

The Tenants API is a readonly discovery service for Agave tenants.
}
\details{
The Tenants API is a readonly discovery service to assist in the location,
configuration, and activation of Agave tenants.  The Tenants API contains
the canonical source of information for tenant baseUrl and identifiers,
which are both used to bootstrap the Agave tooling.
}
\section{Methods}{

\describe{

\code{$getTenantDetails(codeOrUuid)} gets detailed description for an individual
tenant by either code or uuid.

\strong{Usage}\preformatted{tenants$getTenantDetails(codeOrUuid="agave.prod")
}

\code{$listTenants()} List all active tenants with optional pagination;

\strong{Usage}\preformatted{tenants$listTenants()
}

}
}

\seealso{
\link[rAgave:Tenant]{rAgave::Tenant} \link[rAgave:ClientsApi]{rAgave::ClientsApi} \link[rAgave:Client]{rAgave::Client}
}
\keyword{datasets}
