% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Token.r
\docType{data}
\name{Token}
\alias{Token}
\title{Token Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Token
}
\description{
Token Class
}
\section{Fields}{

\describe{
\item{\code{access_token}}{}

\item{\code{refresh_token}}{}

\item{\code{created_at}}{}

\item{\code{expires_in}}{}

\item{\code{expires_at}}{}

\item{\code{username}}{}

\item{\code{password}}{}
}}

\keyword{datasets}
