% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.r
\docType{data}
\name{Transform}
\alias{Transform}
\title{Transform Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Transform
}
\description{
Transform Class
}
\section{Fields}{

\describe{
\item{\code{decoders}}{The available decoders for this transform.}

\item{\code{description}}{Description of this transform.}

\item{\code{descriptionurl}}{The URL to find out more information.}

\item{\code{enabled}}{Whether this transform is enabled or not.}

\item{\code{encoder}}{}

\item{\code{name}}{The name of this transform.}

\item{\code{tags}}{The tags describing this transform.}
}}

\keyword{datasets}
