% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformDecoder.r
\docType{data}
\name{TransformDecoder}
\alias{TransformDecoder}
\title{TransformDecoder Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TransformDecoder
}
\description{
TransformDecoder Class
}
\section{Fields}{

\describe{
\item{\code{description}}{The description of what this decoder is and what it does.}

\item{\code{name}}{The decoder name.}
}}

\keyword{datasets}
