% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformEncoder.r
\docType{data}
\name{TransformEncoder}
\alias{TransformEncoder}
\title{TransformEncoder Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TransformEncoder
}
\description{
TransformEncoder Class
}
\section{Fields}{

\describe{
\item{\code{description}}{The description of what this encoder is and what it does.}

\item{\code{name}}{The encoder name.}
}}

\keyword{datasets}
