% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformRequest.r
\docType{data}
\name{TransformRequest}
\alias{TransformRequest}
\title{TransformRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TransformRequest
}
\description{
TransformRequest Class
}
\section{Fields}{

\describe{
\item{\code{nativeFormat}}{The original file type of the file. If not given, the file type is assumed to be raw.}

\item{\code{url}}{The uri to which the transformed file will be staged.}

\item{\code{callbackUrl}}{The URI to notify when the transfer is complete. This can be an email address or http URL. If a URL is given, a GET will be made to this address. URL templating is supported. Valid template values are: ${NAME}, ${SOURCE_FORMAT}, ${DEST_FORMAT}, ${STATUS}}
}}

\keyword{datasets}
