% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDRequest.r
\docType{data}
\name{UUIDRequest}
\alias{UUIDRequest}
\title{UUIDRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UUIDRequest
}
\description{
Request object to create a new UUID
}
\section{Fields}{

\describe{
\item{\code{type}}{The resource type for which the UUID should be created.}
}}

\keyword{datasets}
