% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDResolution.r
\docType{data}
\name{UUIDResolution}
\alias{UUIDResolution}
\title{UUIDResolution Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UUIDResolution
}
\description{
Resource to which a tag has been associated
}
\section{Fields}{

\describe{
\item{\code{uuid}}{uuid created/requested}

\item{\code{type}}{the resource type represented by this uuid}

\item{\code{representation}}{the full object representation for this uuid}
}}

\keyword{datasets}
