% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UuidsApi.r
\docType{data}
\name{UuidsApi}
\alias{UuidsApi}
\title{Agave Uuids API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
uuids <- UuidsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
UuidsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{createUuid} Generate a new UUID for a given resource type

\strong{Usage}\preformatted{

`resolveUuid` Resolve one or more UUID to their resource URL and type

**Usage**

}

\code{resolveUuids} Resolve one or more UUID to their resource URL and type

\strong{Usage}\preformatted{
}
}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::UUIDRequest][rAgave::UUIDResolution][rAgave::UUIDSummary]
}
\keyword{datasets}
