% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDSummary.r
\docType{data}
\name{UUIDSummary}
\alias{UUIDSummary}
\title{UUIDSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UUIDSummary
}
\description{
Resource to which a tag has been associated
}
\section{Fields}{

\describe{
\item{\code{message}}{success or failure}

\item{\code{result}}{}

\item{\code{status}}{success or failure}

\item{\code{version}}{API version number}
}}

\keyword{datasets}
