% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppsApi.r
\docType{data}
\name{AppsApi}
\alias{AppsApi}
\title{Agave Apps API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
apps <- AppsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
AppsApi Class
}
\details{
An app, in the context of Agave, is an executable code available for invocation through the Agave Jobs service on a specific execution system. Put another way, an app is a piece of code that you can run on a specific system. If a single code needs to be run on multiple systems, each combination of app and system needs to be defined as an app.

Apps are language agnostic and may or may not carry with them their own dependencies. (More on bundling your app in a moment.) Any code that can be forked at the command line or submitted to a batch scheduler can be registered as an Agave app and run through the Jobs service.

The Apps service is the central registry for all Agave apps. The Apps service provides permissions, validation, archiving, and revision information about each app in addition to the usual discovery capability. The rest of this tutorial explains in detail how to register an app to the Apps service, how to manage and share apps, and what the different application scopes mean.

Read more about the Apps API in the \href{http://docs.agaveplatform.org/#/apps}{Agave Developer's Guide}
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{$addApp()} Register and update new applications.

\strong{Usage}\preformatted{apps$addApp$listApps(naked, body, )
}

\code{$addAppPermission()} Grant a user permission for an application.

\strong{Usage}\preformatted{apps$addAppPermission$listApps(appId, body, naked, )
}

\code{$clearAppPermissions()} Deletes all permissions on an application.

\strong{Usage}\preformatted{apps$clearAppPermissions$listApps(appId, naked, )
}

\code{$deleteApp()} Deletes an application.

\strong{Usage}\preformatted{apps$deleteApp$listApps(appId, naked, )
}

\code{$deleteAppPermission()} Deletes all permissions for the given user on an application.

\strong{Usage}\preformatted{apps$deleteAppPermission$listApps(appId, username, naked, )
}

\code{$getAppDetails()} Get details of an application by its unique id.

\strong{Usage}\preformatted{apps$getAppDetails$listApps(appId, naked, search, )
apps$getAppDetails$listApps(search=list("foo"="bar"))
}

\code{$getAppPermission()} Get a specific user permission for an application.

\strong{Usage}\preformatted{apps$getAppPermission$listApps(appId, username, filter, naked, )
}

\code{$getAppSubmissionForm()} Get a submission form for the named application.

\strong{Usage}\preformatted{apps$getAppSubmissionForm$listApps(appId, naked, search, )
apps$getAppSubmissionForm$listApps(search=list("foo"="bar"))
}

\code{$invokeAppAction()} Edit an application.

\strong{Usage}\preformatted{apps$invokeAppAction$listApps(appId, body, naked, )
}

\code{$listAppHistory()} List the event history of this app

\strong{Usage}\preformatted{apps$listAppHistory$listApps(naked, appId, status, created, limit, offset, search, )
apps$listAppHistory$listApps(search=list("foo"="bar"))
}

\code{$listAppPermissions()} Get the permission for this application.

\strong{Usage}\preformatted{apps$listAppPermissions$listApps(appId, filter, naked, limit, offset, search, )
apps$listAppPermissions$listApps(search=list("foo"="bar"))
}

\code{$listApps()} List apps

\strong{Usage}\preformatted{apps$listApps$listApps(public, name, executionSystem, tags, filter, naked, limit, offset, search, )
apps$listApps$listApps(search=list("foo"="bar"))
}

\code{$updateApp()} Update an application.

\strong{Usage}\preformatted{apps$updateApp$listApps(appId, body, naked, )
}

\code{$updateAppPermission()} Add or update a user permission for an application.

\strong{Usage}\preformatted{apps$updateAppPermission$listApps(appId, username, body, naked, )
}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::Application][rAgave::ApplicationAction][rAgave::ApplicationSummary][rAgave::HistoryEvent][rAgave::Permission]
}
\keyword{datasets}
