% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.r
\docType{data}
\name{Client}
\alias{Client}
\title{Client Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Client
}
\description{
Client Class
}
\section{Fields}{

\describe{
\item{\code{key}}{}

\item{\code{secret}}{}

\item{\code{callbackUrl}}{}

\item{\code{description}}{}

\item{\code{clientName}}{}

\item{\code{tier}}{}
}}

\keyword{datasets}
