% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInfo.r
\docType{data}
\name{FileInfo}
\alias{FileInfo}
\title{FileInfo Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileInfo
}
\description{
FileInfo Class
}
\section{Fields}{

\describe{
\item{\code{format}}{The file type of the file.}

\item{\code{lastModified}}{The date this file was last modified in ISO 8601 format.}

\item{\code{length}}{The length of the file/folder.}

\item{\code{mimeType}}{The mime type of the file/folder. If unknown, it defaults to application/binary.}

\item{\code{name}}{The name of the file/folder.}

\item{\code{path}}{The absolute path to the file/folder.}

\item{\code{permissions}}{The remote system permission of the invoking user on the file/folder.}

\item{\code{system}}{The id of the system where this file lives.}

\item{\code{type}}{The filetype. Almost always "raw-0" or "dir"}
}}

\keyword{datasets}
