% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorsApi.r
\docType{data}
\name{MonitorsApi}
\alias{MonitorsApi}
\title{Agave Monitors API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
monitors <- MonitorsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
MonitorsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{addMonitoringTasks} Add a new monitoring task

\strong{Usage}\preformatted{

`deleteMonitoringTask` Deletes a monitor.

**Usage**

}

\code{forceMonitoringTaskCheck} Forces a monitor check to run.

\strong{Usage}\preformatted{

`getMonitoringTask` Retrieve a specific monitor.

**Usage**

}

\code{getMonitoringTaskCheck} Retrieve a specific monitor check

\strong{Usage}\preformatted{

`listMonitoringTaskChecks` Retrieve checks for a specific monitor

**Usage**

}

\code{listMonitoringTasks} Retrieve Monitor for a specific resource.

\strong{Usage}\preformatted{

`updateMonitoringTask` Updates an existing monitor.

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::MonitoringTaskCheck][rAgave::MonitoringTaskDetails][rAgave::MonitoringTaskSummary]
}
\keyword{datasets}
