% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagApi.r
\docType{data}
\name{TagApi}
\alias{TagApi}
\title{Agave Tag API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tag <- TagApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
TagApi Class
}
\details{
Read more about the Tag API in the \href{http://docs.agaveplatform.org/#/tag}{Agave Developer's Guide}
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{$addTag()} Creat a new tag

\strong{Usage}\preformatted{tag$addTag$listTag(body, naked, )
}

\code{$addTagResource()} Add resources to tag

\strong{Usage}\preformatted{tag$addTagResource$listTag(tagId, body, naked, )
}

\code{$clearTagPermissions()} Clear tag permissions

\strong{Usage}\preformatted{tag$clearTagPermissions$listTag(tagId, naked, )
}

\code{$clearTagResources()} Clear all tagged resources

\strong{Usage}\preformatted{tag$clearTagResources$listTag(tagId, naked, )
}

\code{$deleteTag()} Delete tag

\strong{Usage}\preformatted{tag$deleteTag$listTag(tagId, naked, )
}

\code{$deleteTagPermissionForUser()} Remove user tag permission

\strong{Usage}\preformatted{tag$deleteTagPermissionForUser$listTag(tagId, username, naked, )
}

\code{$getTagDetails()} Tag Details

\strong{Usage}\preformatted{tag$getTagDetails$listTag(tagId, naked, search, )
tag$getTagDetails$listTag(search=list("foo"="bar"))
}

\code{$getTagPermissionForUser()} List tag permissions for a user

\strong{Usage}\preformatted{tag$getTagPermissionForUser$listTag(tagId, username, naked, )
}

\code{$getTaggedResource()} Get tagged item details

\strong{Usage}\preformatted{tag$getTaggedResource$listTag(tagId, uuid, naked, )
}

\code{$listTagHistory()} Tag history

\strong{Usage}\preformatted{tag$listTagHistory$listTag(tagId, naked, limit, offset, search, )
tag$listTagHistory$listTag(search=list("foo"="bar"))
}

\code{$listTagPermissions()} List permissionss set for a given tag id

\strong{Usage}\preformatted{tag$listTagPermissions$listTag(tagId, naked, limit, offset, search, )
tag$listTagPermissions$listTag(search=list("foo"="bar"))
}

\code{$listTagResources()} Tagged items

\strong{Usage}\preformatted{tag$listTagResources$listTag(tagId, naked, limit, offset, search, )
tag$listTagResources$listTag(search=list("foo"="bar"))
}

\code{$listTags()} Available tags

\strong{Usage}\preformatted{tag$listTags$listTag(name, resourceType, resourceId, lastModified, created, naked, limit, offset, search, )
tag$listTags$listTag(search=list("foo"="bar"))
}

\code{$removeTagResource()} Remove resource from tag

\strong{Usage}\preformatted{tag$removeTagResource$listTag(tagId, uuid, naked, )
}

\code{$updateTag()} Update tag

\strong{Usage}\preformatted{tag$updateTag$listTag(tagId, body, naked, )
}

\code{$updateTagPermission()} Grant permission

\strong{Usage}\preformatted{tag$updateTagPermission$listTag(tagId, body, naked, )
}

\code{$updateTagPermissionForUser()} Update user permission on tag

\strong{Usage}\preformatted{tag$updateTagPermissionForUser$listTag(tagId, username, body, naked, )
}

\code{$updateTagResource()} Add resource to tag

\strong{Usage}\preformatted{tag$updateTagResource$listTag(tagId, uuid, body, naked, )
}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} [rAgave::Permission][rAgave::Tag][rAgave::TagHistory][rAgave::TagResource][rAgave::TagSummary]
}
\keyword{datasets}
