% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tenant.r
\docType{data}
\name{Tenant}
\alias{Tenant}
\title{Tenant Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tenant
}
\description{
Tenant Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{baseUrl}}{}

\item{\code{code}}{}

\item{\code{created}}{}

\item{\code{lastUpdated}}{}

\item{\code{contact}}{}
}}

\keyword{datasets}
