% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformsApi.r
\docType{data}
\name{TransformsApi}
\alias{TransformsApi}
\title{Agave Transforms API operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
transforms <- TransformsApi$new( apiClient = ApiClient$new(), cache = AgaveCache$new() )
}
\description{
TransformsApi Class
}
\section{Arguments}{

\describe{

\code{apiClient} a \link[rAgave:ApiClient]{rAgave::ApiClient} instance preconfigured to speak to the Agave Platform.

\code{cache} an \link[rAgave:AgaveCache]{rAgave::AgaveCache} instance pointing to the persistent auth cache file on disk.

\code{responseType} the default object type methods of this class should return. Valid values are:
\itemize{
\item \strong{raw}: the raw JSON string response from the API. You may optionally include
\code{pretty=TRUE} in the method call to receive pretty printed JSON.
\item \strong{list}: a standard R list object. This is the default value.
\item \strong{dt}: a data table
}

}
}

\section{Methods}{

\describe{

\code{asyncTransform} Transform a file and stage it to a specified location.

\strong{Usage}\preformatted{

`getTransform` Find the transform of the given uuid

**Usage**

}

\code{listTransforms} List and search for transforms

\strong{Usage}\preformatted{

`syncTransform` Transform a file and download it directly.

**Usage**

}

}
}

\seealso{
\link[rAgave:ApiClient]{rAgave::ApiClient} \link[rAgave:Agave]{rAgave::Agave} \link[rAgave:TransformRequest]{rAgave::Transform}
}
\keyword{datasets}
